Rename `classLsEvent.py` -> `ls_event.py`.
